[HBASE-7424] Enable the DeltaEncoding for the HFileOutputFormat
Author: manukranthk

Summary: HFileOutputFormat has a writer embedded but it is not configured to use the DeltaEncoding and FavoredNodes. This revision is to add that support to the HFileOutputFormat while it is used as an OutputFormat either the Mapper or the Reducer for a MapReduce task.

Test Plan: Adding Unit Tests to test the working of the HFileOutputFormat

Reviewers: liyintang, kannan

CC: hbase-eng@, astewart

Differential Revision: https://phabricator.fb.com/D669389

Task ID: 1964673

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1432708 13f79535-47bb-0310-9956-ffa450edef68