[0.89-fb] [ HBASE-7517] Do not throw PreemptiveFastFailExceptions when caused by another PFFE
Author: aaiyer

Summary:
Sometimes, if we are looking up a region in META, and both META  and
ROOT are unavailable, then the PFFE thrown by root's unavailability seems
to gets caught in the same function and is processed twice. We should avoid
this to prevent the race condition that liyin mentioned.

Test Plan: unit tests

Reviewers: liyintang

Reviewed By: liyintang

CC: hbase-eng@

Differential Revision: https://phabricator.fb.com/D672911

Task ID: 1966872

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1432709 13f79535-47bb-0310-9956-ffa450edef68