[HBASE-7620] Disables load balancer during the batchedUpload test
Author: rshroff

Summary:
The testBatchedUploadWithRegionMoves might fail sometimes if the load
balancer kicks-in during the endBatchedUpload() call. During this call
while fetching the flush times of the regions on the regionservers, if
the region location changes, it throws an exception.

It might be good to make the endBatchedUpload() handle benign region
movements caused by load balancer. If the region has moved, then we can
verify the flush times from the new location. I will file a separate
task for this change. In the mean while, in order to make the test case
pass, I am disabling the load balancer during the write period.

Test Plan:
Exexute the TestBatchedUpload test, it should pass and also
run the MR Unit Test Suite

Reviewers: liyintang, aaiyer

Reviewed By: liyintang

CC: hbase-eng@

Differential Revision: https://phabricator.fb.com/D681445

Task ID: 2028850

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1435307 13f79535-47bb-0310-9956-ffa450edef68