[0.89-fb] [ HBASE-7517] fix PFFE issue when ROOT moves around
Author: aaiyer

Summary:
Fix the issue in TITAN018, that gets appservers
to be stuck throwing PFFE.

It looks like the appservers that are having this trouble
are app servers that have the old-notion of ROOT. This seems
to not get updated, due to the fast fail mechanisms.

Ensure that we always get/refresh the knowledge from ZK, if
the root region server is expected to be inFastFail mode.

Test Plan:
Repro and test.
See if this can be programmtically done as well.

Reviewers: liyintang

Reviewed By: liyintang

CC: hbase-eng@

Differential Revision: https://phabricator.fb.com/D686758

Task ID: 1966872

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1459452 13f79535-47bb-0310-9956-ffa450edef68