[HBASE-7707][89-fb] A simple tool which will COPY or create a smaller table out of another table
Author: adela

Summary:
as in title, 2 possibilities:
1) user can bulkupload the old table to a new table
2) can create smaller table out of a big table by specifying the number of regions (and KVs) which will be copied.
--- HOW TO RUN THIS---
> ./bin/hbase org.apache.hadoop.hbase.mapreduce.TableCopyUtils -bl
Name of the old table: test_table
Name of the new table: abcb
Do you want to set start and stop row for the old table scanner? [y/n] n

(this will create a full copy of table test_table and the new table will be called abcb)

Test Plan:
ran this locally on a very small table. Planning to run this
on tsh25 to create a small table out of cell5_v1 for compaction hook testing.

Reviewers: liyintang

Reviewed By: liyintang

CC: hbase-eng@

Differential Revision: https://phabricator.fb.com/D686059

Task ID: 2051427

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1459454 13f79535-47bb-0310-9956-ffa450edef68