[HBASE-7732] Honors the hfile.io.bytes.per.checksum parameter for HFile Writer
Author: rshroff

Summary:
The HFile Writer used to take the default value
DEFAULT_BYTES_PER_CHECKSUM for bytesPerChecksum. The
hfile.io.bytes.per.checksum was added to provide an option to change
this setting. The HFile Writer create() was not reading this
configuration option. Changed it to read it on every create() call.

Test Plan:
Added a small unit test to make sure that the getBytesPerChecksum
is called and the value from the conf is used.
More suggestions welcomed.

MR:
http://hbaseproxy001.prn1.facebook.com/mr_jt/hbasebackup003-ash3/jobdetails.jsp?jobid=job_201209211223_47382&refresh=30

Reviewers: liyintang

Reviewed By: liyintang

CC: hbase-eng@

Differential Revision: https://phabricator.fb.com/D694332

Task ID: 2064279

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1459456 13f79535-47bb-0310-9956-ffa450edef68