[0.89-fb] [HBASE-7517] Improve the PFFE mechanism such that (i) Ensure that if a thread gets the AtomicBoolean for a single server, it will not be stopped to get the AtomicBoolean for a different server. (ii) PFFE is only thrown for threads that are not a
Author: aaiyer

Summary:
Improvements to the PFFE mechanism, so as to let the
retrying threads to handle errors as if they were not in PFFE mode.

Threads that are allowed to go through (1 per server) will follow the
same logic as normal threads -- when not in fast fail mode. (The only
    exception being that all errors will be logged as LOG.error).
These threads will never throw a PFFE.

Threads that are not allowed to go through, will throw PFFE Exception.

Test Plan:
Add unit test to debug.
Test it out on tsh025.

Reviewers: liyintang, kannan, kranganathan, adela

Reviewed By: liyintang

CC: hbase-eng@

Differential Revision: https://phabricator.fb.com/D690778

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1459464 13f79535-47bb-0310-9956-ffa450edef68