[HBASE-5032] Add Delete Column into the delete bloom filter
Author: adela

Summary:
https://issues.apache.org/jira/browse/HBASE-5032

Previously, the delete family bloom filter only contains the row key
which has the delete family. It helps us to avoid the top-row seek
operation.

This jira attempts to add the delete column into this delete bloom
filter as well (rename the delete family bloom filter as delete bloom
filter).

The motivation is to save seek ops for scan time-range queries if we
know there is no delete column for this row/column.
We can seek directly to the exact timestamp we are interested in,
instead of seeking to the latest timestamp and keep skipping to find
out whether there is any delete column before the interested timestamp.

Test Plan: write unit tests

Reviewers: kannan, liyintang, kranganathan

Reviewed By: liyintang

CC: erling, hbase-eng@

Differential Revision: https://phabricator.fb.com/D622271

Task ID: 849012, 1980066

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1459466 13f79535-47bb-0310-9956-ffa450edef68