[0.89-fb] [HBASE-7852] fix TestLoadIncrementalHFiles
Author: aaiyer

Summary:
TestIncrementalHFiles  does not use the Tool.run() mechanism to
do the bulk load.

In order to fix the test, ensure that we pick up the default options in
the constructor. If Tool.run() is called, the default options are overridden.

Test Plan: TestLoadIncrementalHFiles

Reviewers: liyintang

Reviewed By: liyintang

CC: hbase-eng@

Differential Revision: https://phabricator.fb.com/D718323

Blame Revision: https://phabricator.fb.com/D693206

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1459468 13f79535-47bb-0310-9956-ffa450edef68