[HBASE-8194] Fixing the non deterministic failure of TestHFileOutputFormat.
Author: manukranthk

Summary:
The test creates regions in the meta region and waits for the master to allot them to regionservers. When the favored nodes were set, the timout was not completely reliable to say that they will be set.
Instead setting the favored nodes while creating the regions directly in the starting would ensure that the master creates the regions assigning selected favored nodes.

Test Plan: Unit Tests

Reviewers: liyintang, rshroff, adela

Reviewed By: rshroff

CC: hbase-eng@

Differential Revision: https://phabricator.fb.com/D728493

Task ID: 2138207

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1461253 13f79535-47bb-0310-9956-ffa450edef68