[HBASE-5263] Letting the HFileWriter know, if a compaction is in progress or not
Author: rshroff

Summary:
We don't need to calculate whether the current block should be cached
because of compaction or not if the writer was not created by
compaction.

Also currently, its very difficult from the logs to know if the stats dumped
out for a Writer are for compaction or normal flush.

This change prevents calculating % of cached keys in case the writer is
not being used by compaction.

Test Plan: mr

Reviewers: liyintang, aaiyer, manukranthk

Reviewed By: manukranthk

CC: hbase-eng@

Differential Revision: https://phabricator.fb.com/D760745

Task ID: 2016630

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1464742 13f79535-47bb-0310-9956-ffa450edef68