[HBASE-8371] Removes the dependency on thrift for RegionException
Author: rshroff

Summary:
RegionException was derived from IOError to capture the backoff timeout,
information required by the thrift clients to reduce the load on the
overloaded region servers. The exception was derived from IOError to
make it compatible with thrift. When this change was ported to neptune,
it created a new dependency on the thrift package.

HBase uses version 0.80 while many of the customers either use the
internal thrift or 0.90. This incompatibility is causing problems on the
app server which I think is using internal thrift.

Hence, we decided to get rid of this dependency on the server side by
deriving the RegionException again from IOException and storing the
backOffTime as a class member variable as it was done earlier.

On the thrift server side, when it gets an exception which is an
instance of RegionException, it will fetch the timeBackoff value and
pass it on to the client appropriately.

Test Plan:
I will test this on SEARCHHBASE004-FRC1 as they have a c++
thrift client communicating with HBase

Reviewers: liyintang, aaiyer, avf, adela

Reviewed By: adela

CC: hbase-eng@, nathan, davejwatson

Differential Revision: https://phabricator.fb.com/D779316

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1470794 13f79535-47bb-0310-9956-ffa450edef68