[HBASE-8467] Adds an option to blacklist a server for RollingRestart
Author: rshroff

Summary:
Following are the master side changes to allow a region server to be
blacklisted. By blacklisted, it means that the HMaster will not assign
new regions to this region server. This will help us in RollingRestart
where we want to drain the region from a particular region server.

It also provides knobs to clear the entire list for recovery purposes.

Test Plan:
Unit test and mr
One test is failing (Investigating)

Reviewers: aaiyer, liyintang

Reviewed By: aaiyer

CC: hbase-eng@, paultuckfield, amit

Differential Revision: https://phabricator.fb.com/D781160

Task ID: 2229110

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1477750 13f79535-47bb-0310-9956-ffa450edef68