[HBASE-8083] Fixes TestHCM
Author: rshroff

Summary:
One of the test cases testRemoteServerFailure, changed the HRegion class
used to perform the operations. The test did not reset the class back to
original one after completion. This would cause errors to any other
tests executing after this test case.

Changed it to reset the class to the original one on completion. Also,
changed the order to ensure that there are test cases which run after
this test case to check whether its works or not.

Test Plan: mr

Reviewers: liyintang, manukranthk

Reviewed By: liyintang

CC: hbase-eng@

Differential Revision: https://phabricator.fb.com/D809815

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1483524 13f79535-47bb-0310-9956-ffa450edef68